Move type serialization functions to a separate type_serialization.h (#9537)
