+++ ydb/core/tx/datashard/export_common.h
@@ -8,8 +8,6 @@
 
 #include <util/generic/map.h>
 #include <util/generic/maybe.h>
-#include <util/generic/string.h>
-#include <util/generic/vector.h>
 
 #if defined EXPORT_LOG_T || \
     defined EXPORT_LOG_D || \
@@ -29,19 +27,11 @@
 #define EXPORT_LOG_E(stream) LOG_ERROR_S(*TlsActivationContext, NKikimrServices::DATASHARD_BACKUP, "[Export] [" << LogPrefix() << "] " << stream)
 #define EXPORT_LOG_C(stream) LOG_CRIT_S(*TlsActivationContext, NKikimrServices::DATASHARD_BACKUP, "[Export] [" << LogPrefix() << "] " << stream)
 
-namespace NKikimr {
-namespace NDataShard {
+namespace NKikimr::NDataShard {
 
 TMaybe<Ydb::Table::CreateTableRequest> GenYdbScheme(
     const TMap<ui32, TUserTable::TUserColumn>& columns,
     const NKikimrSchemeOp::TPathDescription& pathDesc);
 
-TString DecimalToString(const std::pair<ui64, i64>& loHi, const NScheme::TTypeInfo& typeInfo);
-TString DyNumberToString(TStringBuf data);
-bool DecimalToStream(const std::pair<ui64, i64>& loHi, IOutputStream& out, TString& err, const NScheme::TTypeInfo& typeInfo);
-bool DyNumberToStream(TStringBuf data, IOutputStream& out, TString& err);
-bool PgToStream(TStringBuf data, const NScheme::TTypeInfo& typeInfo, IOutputStream& out, TString& err);
-bool UuidToStream(const std::pair<ui64, ui64>& loHi, IOutputStream& out, TString& err);
 
-} // NDataShard
-} // NKikimr
+} // NKikimr::NDataShard
