+++ ydb/core/tx/datashard/ya.make
@@ -202,6 +202,7 @@ SRCS(
     store_distributed_erase_tx_unit.cpp
     store_scheme_tx_unit.cpp
     store_snapshot_tx_unit.cpp
+    type_serialization.cpp
     volatile_tx.cpp
     wait_for_plan_unit.cpp
     wait_for_stream_clearance_unit.cpp
